SharedPage.js migration
